fix: change vtable-plugins files
