+++ packages/vtable-plugins/src/index.ts
@@ -1,5 +1,3 @@
 export * from './carousel-animation';
 export * from './invert-highlight';
-export * from './list-tree-stick-cell';
-export * from './custom-cell-style';
 export * from './header-highlight';
