+++ packages/vtable/src/plugins/list-tree-stick-cell.ts
@@ -1,10 +1,15 @@
 import { isArray, isValid } from '@visactor/vutils';
-import type { ListTable } from '@visactor/vtable/es/ListTable';
-import { Factory } from '@visactor/vtable/es/core/factory';
-import { Group } from '@visactor/vtable/es/scenegraph/graphic/group';
-import { createRect, type Graphic } from '@visactor/vtable/es/vrender';
-import { updateCell } from '@visactor/vtable/es/scenegraph/group-creater/cell-helper';
-import { getTargetCell } from '@visactor/vtable/es/event/util';
+import type { ListTable } from '../ListTable';
+import { Group } from '../scenegraph/graphic/group';
+import { updateCell } from '../scenegraph/group-creater/cell-helper';
+import type { Graphic } from '@src/vrender';
+import { createRect } from '@src/vrender';
+import { Factory } from '../core/factory';
+import { getTargetCell } from '../event/util';
+
+export interface IListTreeStickCellPlugin {
+  new (table: ListTable): ListTreeStickCellPlugin;
+}
 
 export class ListTreeStickCellPlugin {
   table: ListTable;
@@ -292,4 +297,4 @@ export const registerListTreeStickCellPlugin = () => {
   Factory.registerComponent('listTreeStickCellPlugin', ListTreeStickCellPlugin);
 };
 
-export type IListTreeStickCellPlugin = typeof ListTreeStickCellPlugin;
+// export type IListTreeStickCellPlugin = typeof ListTreeStickCellPlugin;
